+++ TP_PAV_3k2/TP_PAV_3k2/Formularios/Producto/AltaProductos.cs
@@ -172,8 +172,8 @@ namespace TP_PAV_3k2.Formularios.Producto
                 }
                 else
                 {
-                    float defaul;
-                    if (float.TryParse(txtPrecioCompra.Text, out defaul))
+                    decimal defaul;
+                    if (decimal.TryParse(txtPrecioCompra.Text, out defaul))
                     {
                         producto.PrecioCompra = defaul;
                     }
@@ -202,8 +202,8 @@ namespace TP_PAV_3k2.Formularios.Producto
                 }
                 else
                 {
-                    float defaul;
-                    if (float.TryParse(txtPrecioVenta.Text, out defaul))
+                    decimal defaul;
+                    if (decimal.TryParse(txtPrecioVenta.Text, out defaul))
                     {
                         producto.PrecioVenta = defaul;
                     }
@@ -298,8 +298,8 @@ namespace TP_PAV_3k2.Formularios.Producto
                 }
                 else
                 {
-                    float defaul;
-                    if (float.TryParse(txtPrecioCompra.Text, out defaul))
+                    decimal defaul;
+                    if (decimal.TryParse(txtPrecioCompra.Text, out defaul))
                     {
                         producto.PrecioCompra = defaul;
                     }
@@ -328,8 +328,8 @@ namespace TP_PAV_3k2.Formularios.Producto
                 }
                 else
                 {
-                    float defaul;
-                    if (float.TryParse(txtPrecioVenta.Text, out defaul))
+                    decimal defaul;
+                    if (decimal.TryParse(txtPrecioVenta.Text, out defaul))
                     {
                         producto.PrecioVenta = defaul;
                     }
